requirements.txt
appdirs==1.4.4
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
black==21.5b2
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
decorator==5.1.1
Deprecated==1.2.13
Django==4.0.5
django-cors-headers==3.11.0
django-environ==0.4.5
django-ninja==0.16.1
django-stubs==1.9.0
django-stubs-ext==0.3.1
executing==0.8.3
flake8==3.9.2
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.1.0
huey==2.4.1
importlib-metadata==4.4.0
ipython==8.3.0
isort==5.8.0
jedi==0.18.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
nameparser==1.0.6
parso==0.8.3
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
Pygments==2.12.0
PyJWT==2.3.0
pytz==2021.1
redis==4.0.1
regex==2021.4.4
six==1.16.0
sqlparse==0.4.1
stack-data==0.2.0
toml==0.10.2
traitlets==5.1.1
typed-ast==1.4.3
types-pytz==2021.3.0
types-PyYAML==6.0.1
typing-extensions==3.10.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
